+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Improved XMPP vCard data parsing
 ### Fixed
 - Added missing user-agent headers
+### Removed
+- jsdom dependency
 
 ## [0.18.0] - 2022-11-17
 ### Changed